+++ src/Vehicle/Vehicle.h
@@ -633,7 +633,7 @@ public:
     typedef enum {
         MavCmdResultCommandResultOnly,          ///< commandResult specifies full success/fail info
         MavCmdResultFailureNoResponseToCommand, ///< No response from vehicle to command
-        MavCmdResultFailureDuplicateCommand,    ///< Unabled to send command since duplicate is already being waited on for response
+        MavCmdResultFailureDuplicateCommand,    ///< Unable to send command since duplicate is already being waited on for response
     } MavCmdResultFailureCode_t;
 
     /// Callback for sendMavCommandWithHandler