+++ src/comm/MockLink.cc
@@ -1567,13 +1567,15 @@ bool MockLink::_handleRequestMessage(const mavlink_command_long_t& request, bool
 
     switch ((int)request.param1) {
     case MAVLINK_MSG_ID_COMPONENT_INFORMATION:
-        switch (static_cast<int>(request.param2)) {
-        case COMP_METADATA_TYPE_VERSION:
-            _sendVersionMetaData();
-            return true;
-        case COMP_METADATA_TYPE_PARAMETER:
-            _sendParameterMetaData();
-            return true;
+        if (_firmwareType == MAV_AUTOPILOT_PX4) {
+            switch (static_cast<int>(request.param2)) {
+            case COMP_METADATA_TYPE_VERSION:
+                _sendVersionMetaData();
+                return true;
+            case COMP_METADATA_TYPE_PARAMETER:
+                _sendParameterMetaData();
+                return true;
+            }
         }
         break;
     case MAVLINK_MSG_ID_DEBUG: